chore: make requested changes

+++ src/database.ts
@@ -26,25 +26,33 @@ export async function setUp(): Promise<void> {
 }
 
 /**
- * Check if the user exists in the database. If not, add them. And if they do, return a boolean
- * indicating if they can play or not
- * @param id The user's ID
+ * Indicates whether this user is registered in the database
+ * @param id The user id to check
+ * @returns {Promise<boolean | undefined>} Whether the user is registered
  */
-export async function checkOrRegisterUser(id: string): Promise<boolean | void> {
+export async function isUserInDB(id: string): Promise<boolean | undefined> {
 	const userRow: QueryResult<UserRow> = await client.query(`SELECT id, status FROM users WHERE id = '${id}'`);
 
-	if (userRow.rowCount === 0) {
-		await client.query(`INSERT INTO users (id, status) VALUES ('${id}', false)`);
-	} else if (userRow.rows[0].status === true) {
+	if (!userRow.rowCount) {
+		return false;
+	} else if (userRow.rows[0].status) {
 		return true;
 	}
 }
 
 /**
- * Used when the user finished the game, making their status true
- * @param id The user's ID
+ * Registers a user in the database
+ * @param id The id of the user to register
  */
-export function played(id: string): void {
+export function registerUser(id: string): void {
+	client.query(`INSERT INTO users (id, status) VALUES ('${id}', false)`);
+}
+
+/**
+ * Sets the status of a user to false in the database
+ * @param id The user id to set as played
+ */
+export function setPlayed(id: string): void {
 	client.query(`UPDATE users SET status = true WHERE id = '${id}'`);
 }
 
@@ -65,9 +73,10 @@ export async function resetUser(id: string): Promise<boolean> {
 }
 
 /**
- * Each day at midnight, this function is called to delete all users and use other word as correct
+ * Deletes/replaces the word in the database to a new one be used and resets all the users
+ * @param replace Wheter to replace the word
  */
-export async function newDay(): Promise<void> {
+export async function newDay(replace = false): Promise<void> {
 	const words: string[] = fs.readFileSync('src/words/wordsList.txt', 'utf8').split('\n');
 	const rw: string = words[Math.floor(Math.random() * words.length)].replace('\r', '');
 
@@ -75,27 +84,12 @@ export async function newDay(): Promise<void> {
 		const word: QueryResult<WordRow> = await client.query(`SELECT word, status FROM words WHERE word = '${rw}'`);
 
 		if (word.rowCount === 0) {
-			await client.query('UPDATE words SET status = false WHERE status = true');
-			await client.query(`INSERT INTO words (word, status) VALUES ('${rw}', true)`);
-			await client.query('DELETE FROM users');
-		} else {
-			continue;
-		}
-	}
-}
+			if (replace) {
+				await client.query('DELETE FROM words WHERE status = true');
+			} else {
+				await client.query('UPDATE words SET status = false WHERE status = true');
+			}
 
-/**
- * Creates a new word that will be used in the game
- */
-export async function newWord(): Promise<void> {
-	const words: string[] = fs.readFileSync('src/words/wordsList.txt', 'utf8').split('\n');
-	const rw: string = words[Math.floor(Math.random() * words.length)].replace('\r', '');
-
-	for (let i = 0; i < words.length; i++) {
-		const word: QueryResult<WordRow> = await client.query(`SELECT word, status FROM words WHERE word = '${rw}'`);
-
-		if (word.rowCount === 0) {
-			await client.query('DELETE FROM words WHERE status = true');
 			await client.query(`INSERT INTO words (word, status) VALUES ('${rw}', true)`);
 			await client.query('DELETE FROM users');
 		} else {
